refactor NoteListView and update test target iOS version
